move some integration tests into unit
