+++ library/src/androidTest/java/org/xmtp/android/library/InstrumentedTest.kt
@@ -1,30 +1,19 @@
 package org.xmtp.android.library
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.filters.FlakyTest
 import kotlinx.coroutines.runBlocking
 import org.junit.Assert.assertEquals
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.xmtp.android.library.messages.Envelope
-import org.xmtp.android.library.messages.InvitationV1
-import org.xmtp.android.library.messages.InvitationV1ContextBuilder
 import org.xmtp.android.library.messages.PrivateKey
 import org.xmtp.android.library.messages.PrivateKeyBuilder
 import org.xmtp.android.library.messages.PrivateKeyBundleBuilder
-import org.xmtp.android.library.messages.SealedInvitationBuilder
-import org.xmtp.android.library.messages.SealedInvitationHeaderV1
 import org.xmtp.android.library.messages.Topic
-import org.xmtp.android.library.messages.createRandom
 import org.xmtp.android.library.messages.encrypted
 import org.xmtp.android.library.messages.generate
-import org.xmtp.android.library.messages.header
-import org.xmtp.android.library.messages.recoverWalletSignerPublicKey
 import org.xmtp.android.library.messages.secp256K1Uncompressed
-import org.xmtp.android.library.messages.toSignedPublicKeyBundle
-import org.xmtp.android.library.messages.toV2
 import org.xmtp.android.library.messages.walletAddress
-import org.xmtp.proto.message.contents.Invitation
 import org.xmtp.proto.message.contents.PrivateKeyOuterClass
 import java.util.Date
 
@@ -74,7 +63,6 @@ class InstrumentedTest {
     }
 
     @Test
-    @FlakyTest
     fun testPublishingAndFetchingContactBundlesWithSavedKeys() {
         val aliceWallet = PrivateKeyBuilder()
         val alice = PrivateKeyOuterClass.PrivateKeyBundleV1.newBuilder().build()
@@ -108,91 +96,5 @@ class InstrumentedTest {
         )
         assert(contact!!.v2.keyBundle.identityKey.hasSignature())
         assert(contact.v2.keyBundle.preKey.hasSignature())
-    }
-
-    @Test
-    fun testEndToEndConversation() {
-        val options =
-            ClientOptions(api = ClientOptions.Api(env = XMTPEnvironment.LOCAL, isSecure = false))
-        val fakeContactWallet = PrivateKeyBuilder()
-        val fakeContactClient = Client().create(account = fakeContactWallet, options = options)
-        fakeContactClient.publishUserContact()
-        val fakeWallet = PrivateKeyBuilder()
-        val client = Client().create(account = fakeWallet, options = options)
-        val contact = client.getUserContact(peerAddress = fakeContactWallet.address)!!
-        assertEquals(contact.walletAddress, fakeContactWallet.address)
-        val created = Date()
-        val invitationContext = Invitation.InvitationV1.Context.newBuilder().also {
-            it.conversationId = "https://example.com/1"
-        }.build()
-        val invitationv1 =
-            InvitationV1.newBuilder().build().createRandom(context = invitationContext)
-        val senderBundle = client.privateKeyBundleV1?.toV2()
-        assertEquals(
-            senderBundle?.identityKey?.publicKey?.recoverWalletSignerPublicKey()?.walletAddress,
-            fakeWallet.address
-        )
-        val invitation = SealedInvitationBuilder.buildFromV1(
-            sender = client.privateKeyBundleV1!!.toV2(),
-            recipient = contact.toSignedPublicKeyBundle(),
-            created = created,
-            invitation = invitationv1
-        )
-        val inviteHeader = invitation.v1.header
-        assertEquals(inviteHeader.sender.walletAddress, fakeWallet.address)
-        assertEquals(inviteHeader.recipient.walletAddress, fakeContactWallet.address)
-        val header = SealedInvitationHeaderV1.parseFrom(invitation.v1.headerBytes)
-        val conversation =
-            ConversationV2.create(client = client, invitation = invitationv1, header = header)
-        assertEquals(fakeContactWallet.address, conversation.peerAddress)
-
-        conversation.send(content = "hello world")
-
-        val conversationList = client.conversations.list()
-        val recipientConversation = conversationList.lastOrNull()
-
-        val messages = recipientConversation?.messages()
-        val message = messages?.firstOrNull()
-        if (message != null) {
-            assertEquals("hello world", message.body)
-        }
-    }
-
-    @Test
-    @FlakyTest
-    fun testCanReceiveV1MessagesFromJS() {
-        val wallet = TestHelpers.FakeWallet.generate()
-        val options =
-            ClientOptions(api = ClientOptions.Api(env = XMTPEnvironment.LOCAL, isSecure = false))
-        val client = Client().create(account = wallet, options = options)
-        val convo = ConversationV1(
-            client = client,
-            peerAddress = "0xf4BF19Ed562651837bc11ff975472ABd239D35B5",
-            sentAt = Date()
-        )
-        convo.send(text = "hello from kotlin")
-        Thread.sleep(1_000)
-        val messages = convo.messages()
-        assertEquals(1, messages.size)
-        assertEquals("hello from kotlin", messages[0].body)
-    }
-
-    @Test
-    @FlakyTest
-    fun testCanReceiveV2MessagesFromJS() {
-        val wallet = PrivateKeyBuilder()
-        val options =
-            ClientOptions(api = ClientOptions.Api(env = XMTPEnvironment.LOCAL, isSecure = false))
-        val client = Client().create(account = wallet, options = options)
-        val convo = client.conversations.newConversation(
-            "0xf4BF19Ed562651837bc11ff975472ABd239D35B5",
-            InvitationV1ContextBuilder.buildFromConversation("https://example.com/4")
-        )
-
-        convo.send(content = "hello from kotlin")
-        Thread.sleep(1_000)
-        val messages = convo.messages()
-        assertEquals(1, messages.size)
-        assertEquals("hello from kotlin", messages[0].body)
     }
 }
