+++ library/src/test/java/org/xmtp/android/library/ConversationTest.kt
@@ -9,6 +9,7 @@ import org.xmtp.android.library.codecs.TextCodec
 import org.xmtp.android.library.messages.ContactBundle
 import org.xmtp.android.library.messages.Envelope
 import org.xmtp.android.library.messages.EnvelopeBuilder
+import org.xmtp.android.library.messages.InvitationV1
 import org.xmtp.android.library.messages.InvitationV1ContextBuilder
 import org.xmtp.android.library.messages.MessageBuilder
 import org.xmtp.android.library.messages.MessageHeaderV2Builder
@@ -16,13 +17,20 @@ import org.xmtp.android.library.messages.MessageV1Builder
 import org.xmtp.android.library.messages.MessageV2Builder
 import org.xmtp.android.library.messages.PrivateKey
 import org.xmtp.android.library.messages.PrivateKeyBuilder
+import org.xmtp.android.library.messages.SealedInvitationBuilder
+import org.xmtp.android.library.messages.SealedInvitationHeaderV1
 import org.xmtp.android.library.messages.SignedContentBuilder
 import org.xmtp.android.library.messages.Topic
+import org.xmtp.android.library.messages.createRandom
 import org.xmtp.android.library.messages.getPublicKeyBundle
+import org.xmtp.android.library.messages.header
+import org.xmtp.android.library.messages.recoverWalletSignerPublicKey
 import org.xmtp.android.library.messages.sign
 import org.xmtp.android.library.messages.toPublicKeyBundle
+import org.xmtp.android.library.messages.toSignedPublicKeyBundle
 import org.xmtp.android.library.messages.toV2
 import org.xmtp.android.library.messages.walletAddress
+import org.xmtp.proto.message.contents.Invitation
 import java.util.Date
 
 class ConversationTest {
@@ -308,5 +316,51 @@ class ConversationTest {
         assertEquals(1, messages.size)
         assertEquals(MutableList(1000) { "A" }.toString(), messages[0].body)
         assertEquals(bobWallet.address, messages[0].senderAddress)
+    }
+
+    @Test
+    fun testEndToEndConversation() {
+        val fakeContactWallet = PrivateKeyBuilder()
+        val fakeContactClient = Client().create(account = fakeContactWallet)
+        fakeContactClient.publishUserContact()
+        val fakeWallet = PrivateKeyBuilder()
+        val client = Client().create(account = fakeWallet)
+        val contact = client.getUserContact(peerAddress = fakeContactWallet.address)!!
+        assertEquals(contact.walletAddress, fakeContactWallet.address)
+        val created = Date()
+        val invitationContext = Invitation.InvitationV1.Context.newBuilder().also {
+            it.conversationId = "https://example.com/1"
+        }.build()
+        val invitationv1 =
+            InvitationV1.newBuilder().build().createRandom(context = invitationContext)
+        val senderBundle = client.privateKeyBundleV1?.toV2()
+        assertEquals(
+            senderBundle?.identityKey?.publicKey?.recoverWalletSignerPublicKey()?.walletAddress,
+            fakeWallet.address
+        )
+        val invitation = SealedInvitationBuilder.buildFromV1(
+            sender = client.privateKeyBundleV1!!.toV2(),
+            recipient = contact.toSignedPublicKeyBundle(),
+            created = created,
+            invitation = invitationv1
+        )
+        val inviteHeader = invitation.v1.header
+        assertEquals(inviteHeader.sender.walletAddress, fakeWallet.address)
+        assertEquals(inviteHeader.recipient.walletAddress, fakeContactWallet.address)
+        val header = SealedInvitationHeaderV1.parseFrom(invitation.v1.headerBytes)
+        val conversation =
+            ConversationV2.create(client = client, invitation = invitationv1, header = header)
+        assertEquals(fakeContactWallet.address, conversation.peerAddress)
+
+        conversation.send(content = "hello world")
+
+        val conversationList = client.conversations.list()
+        val recipientConversation = conversationList.lastOrNull()
+
+        val messages = recipientConversation?.messages()
+        val message = messages?.firstOrNull()
+        if (message != null) {
+            assertEquals("hello world", message.body)
+        }
     }
 }
